+++ pulp_rpm/app/viewsets/__init__.py
@@ -1,0 +1,22 @@
+from .acs import RpmAlternateContentSourceViewSet  # noqa
+from .advisory import UpdateRecordViewSet  # noqa
+from .comps import (  # noqa
+    CompsXmlViewSet,
+    PackageGroupViewSet,
+    PackageCategoryViewSet,
+    PackageEnvironmentViewSet,
+    PackageLangpacksViewSet,
+)
+from .custom_metadata import RepoMetadataFileViewSet  # noqa
+from .distribution import DistributionTreeViewSet  # noqa
+from .modulemd import ModulemdViewSet, ModulemdDefaultsViewSet  # noqa
+from .package import PackageViewSet  # noqa
+from .repository import (  # noqa
+    RpmRepositoryViewSet,
+    RpmRepositoryVersionViewSet,
+    RpmRemoteViewSet,
+    UlnRemoteViewSet,
+    RpmPublicationViewSet,
+    RpmDistributionViewSet,
+    CopyViewSet,
+)
