+++ pulp_rpm/app/viewsets/distribution.py
@@ -1,0 +1,31 @@
+from pulpcore.plugin.viewsets import (
+    ReadOnlyContentViewSet,
+)
+
+from pulp_rpm.app.models import (
+    DistributionTree,
+)
+from pulp_rpm.app.serializers import (
+    DistributionTreeSerializer,
+)
+
+
+class DistributionTreeViewSet(ReadOnlyContentViewSet):
+    """
+    Distribution Tree Viewset.
+
+    """
+
+    endpoint_name = "distribution_trees"
+    queryset = DistributionTree.objects.all()
+    serializer_class = DistributionTreeSerializer
+
+    DEFAULT_ACCESS_POLICY = {
+        "statements": [
+            {
+                "action": ["list", "retrieve"],
+                "principal": "authenticated",
+                "effect": "allow",
+            },
+        ],
+    }
