+++ pulp_rpm/app/viewsets/repository.py
@@ -1,136 +1,46 @@
-import os
 from gettext import gettext as _
 import logging
 
-from django_filters import CharFilter
-from django.utils.timezone import now
 from drf_spectacular.utils import extend_schema
 from rest_framework import viewsets
 from rest_framework.decorators import action
-from rest_framework.generics import get_object_or_404
 from rest_framework.serializers import ValidationError as DRFValidationError
 
-from pulpcore.plugin.models import PulpTemporaryFile
 from pulpcore.plugin.actions import ModifyRepositoryActionMixin
-from pulpcore.plugin.models import AlternateContentSourcePath, RepositoryVersion, TaskGroup
+from pulpcore.plugin.models import RepositoryVersion
 from pulpcore.plugin.tasking import dispatch
 from pulpcore.plugin.serializers import (
     AsyncOperationResponseSerializer,
-    TaskGroupOperationResponseSerializer,
 )
 from pulpcore.plugin.viewsets import (
-    AlternateContentSourceViewSet,
-    ContentFilter,
     DistributionViewSet,
     NamedModelViewSet,
-    NoArtifactContentUploadViewSet,
     OperationPostponedResponse,
     PublicationViewSet,
-    ReadOnlyContentViewSet,
     RemoteViewSet,
     RepositoryVersionViewSet,
     RepositoryViewSet,
     RolesMixin,
-    SingleArtifactContentUploadViewSet,
-    TaskGroupOperationResponse,
 )
 
 from pulp_rpm.app import tasks
 from pulp_rpm.app.constants import SYNC_POLICIES
 from pulp_rpm.app.models import (
-    DistributionTree,
-    Modulemd,
-    ModulemdDefaults,
-    Package,
-    PackageCategory,
-    PackageEnvironment,
-    PackageGroup,
-    PackageLangpacks,
-    RepoMetadataFile,
-    RpmAlternateContentSource,
     RpmDistribution,
     RpmPublication,
     RpmRemote,
     RpmRepository,
     UlnRemote,
-    UpdateRecord,
 )
 from pulp_rpm.app.serializers import (
-    CompsXmlSerializer,
     CopySerializer,
-    DistributionTreeSerializer,
-    MinimalPackageSerializer,
-    MinimalUpdateRecordSerializer,
-    ModulemdDefaultsSerializer,
-    ModulemdSerializer,
-    PackageCategorySerializer,
-    PackageEnvironmentSerializer,
-    PackageGroupSerializer,
-    PackageLangpacksSerializer,
-    PackageSerializer,
-    RepoMetadataFileSerializer,
-    RpmAlternateContentSourceSerializer,
     RpmDistributionSerializer,
     RpmPublicationSerializer,
     RpmRemoteSerializer,
     RpmRepositorySerializer,
     RpmRepositorySyncURLSerializer,
     UlnRemoteSerializer,
-    UpdateRecordSerializer,
 )
-
-
-class PackageFilter(ContentFilter):
-    """
-    FilterSet for Package.
-    """
-
-    sha256 = CharFilter(field_name="_artifacts__sha256")
-
-    class Meta:
-        model = Package
-        fields = {
-            "name": ["exact", "in", "ne"],
-            "epoch": ["exact", "in", "ne"],
-            "version": ["exact", "in", "ne"],
-            "release": ["exact", "in", "ne"],
-            "arch": ["exact", "in", "ne"],
-            "pkgId": ["exact", "in"],
-            "checksum_type": ["exact", "in", "ne"],
-        }
-
-
-class PackageViewSet(SingleArtifactContentUploadViewSet):
-    """
-    A ViewSet for Package.
-
-    Define endpoint name which will appear in the API endpoint for this content type.
-    For example::
-        http://pulp.example.com/pulp/api/v3/content/rpm/packages/
-
-    Also specify queryset and serializer for Package.
-    """
-
-    endpoint_name = "packages"
-    queryset = Package.objects.prefetch_related("_artifacts")
-    serializer_class = PackageSerializer
-    minimal_serializer_class = MinimalPackageSerializer
-    filterset_class = PackageFilter
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-            {
-                "action": ["create"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
 
 
 class RpmRepositoryViewSet(RepositoryViewSet, ModifyRepositoryActionMixin, RolesMixin):
@@ -536,55 +446,6 @@ class UlnRemoteViewSet(RemoteViewSet, RolesMixin):
     }
 
 
-class UpdateRecordFilter(ContentFilter):
-    """
-    FilterSet for UpdateRecord.
-    """
-
-    class Meta:
-        model = UpdateRecord
-        fields = {
-            "id": ["exact", "in"],
-            "status": ["exact", "in", "ne"],
-            "severity": ["exact", "in", "ne"],
-            "type": ["exact", "in", "ne"],
-        }
-
-
-class UpdateRecordViewSet(NoArtifactContentUploadViewSet):
-    """
-    A ViewSet for UpdateRecord.
-
-    Define endpoint name which will appear in the API endpoint for this content type.
-    For example::
-        http://pulp.example.com/pulp/api/v3/content/rpm/advisories/
-
-    Also specify queryset and serializer for UpdateRecord.
-    """
-
-    endpoint_name = "advisories"
-    queryset = UpdateRecord.objects.all()
-    serializer_class = UpdateRecordSerializer
-    minimal_serializer_class = MinimalUpdateRecordSerializer
-    filterset_class = UpdateRecordFilter
-
-    # TODO: adjust this policy after upload access policy design done and in place
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-            {
-                "action": ["create"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
 class RpmPublicationViewSet(PublicationViewSet, RolesMixin):
     """
     ViewSet for Rpm Publications.
@@ -885,397 +746,3 @@ class CopyViewSet(viewsets.ViewSet):
             result.append(r)
 
         return result, shared_repos, exclusive_repos
-
-
-class PackageGroupViewSet(ReadOnlyContentViewSet):
-    """
-    PackageGroup ViewSet.
-    """
-
-    endpoint_name = "packagegroups"
-    queryset = PackageGroup.objects.all()
-    serializer_class = PackageGroupSerializer
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class PackageCategoryViewSet(ReadOnlyContentViewSet):
-    """
-    PackageCategory ViewSet.
-    """
-
-    endpoint_name = "packagecategories"
-    queryset = PackageCategory.objects.all()
-    serializer_class = PackageCategorySerializer
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class PackageEnvironmentViewSet(ReadOnlyContentViewSet):
-    """
-    PackageEnvironment ViewSet.
-    """
-
-    endpoint_name = "packageenvironments"
-    queryset = PackageEnvironment.objects.all()
-    serializer_class = PackageEnvironmentSerializer
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class PackageLangpacksViewSet(ReadOnlyContentViewSet):
-    """
-    PackageLangpacks ViewSet.
-    """
-
-    endpoint_name = "packagelangpacks"
-    queryset = PackageLangpacks.objects.all()
-    serializer_class = PackageLangpacksSerializer
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class DistributionTreeViewSet(ReadOnlyContentViewSet):
-    """
-    Distribution Tree Viewset.
-
-    """
-
-    endpoint_name = "distribution_trees"
-    queryset = DistributionTree.objects.all()
-    serializer_class = DistributionTreeSerializer
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class RepoMetadataFileViewSet(ReadOnlyContentViewSet):
-    """
-    RepoMetadataFile Viewset.
-
-    """
-
-    endpoint_name = "repo_metadata_files"
-    queryset = RepoMetadataFile.objects.all()
-    serializer_class = RepoMetadataFileSerializer
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class ModulemdFilter(ContentFilter):
-    """
-    FilterSet for Modulemd.
-    """
-
-    sha256 = CharFilter(field_name="_artifacts__sha256")
-
-    class Meta:
-        model = Modulemd
-        fields = {
-            "name": ["exact", "in"],
-            "stream": ["exact", "in"],
-        }
-
-
-class ModulemdViewSet(SingleArtifactContentUploadViewSet):
-    """
-    ViewSet for Modulemd.
-    """
-
-    endpoint_name = "modulemds"
-    queryset = Modulemd.objects.all()
-    serializer_class = ModulemdSerializer
-    filterset_class = ModulemdFilter
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class ModulemdDefaultsFilter(ContentFilter):
-    """
-    FilterSet for ModulemdDefaults.
-    """
-
-    sha256 = CharFilter(field_name="_artifacts__sha256")
-
-    class Meta:
-        model = ModulemdDefaults
-        fields = {
-            "module": ["exact", "in"],
-            "stream": ["exact", "in"],
-        }
-
-
-class ModulemdDefaultsViewSet(SingleArtifactContentUploadViewSet):
-    """
-    ViewSet for Modulemd.
-    """
-
-    endpoint_name = "modulemd_defaults"
-    queryset = ModulemdDefaults.objects.all()
-    serializer_class = ModulemdDefaultsSerializer
-    filterset_class = ModulemdDefaultsFilter
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list", "retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-            },
-        ],
-    }
-
-
-class CompsXmlViewSet(viewsets.ViewSet):
-    """
-    ViewSet for comps.xml Upload.
-    """
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["create"],
-                "principal": "authenticated",
-                "effect": "allow",
-                "condition": [
-                    "has_repo_attr_model_or_obj_perms:rpm.modify_content_rpmrepository",
-                ],
-            },
-        ],
-    }
-
-    @extend_schema(
-        description="Trigger an asynchronous task to upload a comps.xml file.",
-        summary="Upload comps.xml",
-        operation_id="rpm_comps_upload",
-        request=CompsXmlSerializer,
-        responses={202: AsyncOperationResponseSerializer},
-    )
-    def create(self, request):
-        """Upload a comps.xml file and create Content from it."""
-        serializer = CompsXmlSerializer(data=request.data, context={"request": request})
-        serializer.is_valid(raise_exception=True)
-
-        # Store TemporaryUpload as a file we can find/use from our task
-        task_payload = {k: v for k, v in request.data.items()}
-        file_content = task_payload.pop("file", None)
-        temp_file = PulpTemporaryFile.init_and_validate(file_content)
-        temp_file.save()
-
-        # Lock destination-repo if we are given one so two uploads can't collide
-        repository = serializer.validated_data.get("repository", None)
-        repo_pk = str(repository.pk) if repository else None
-        replace = serializer.validated_data.get("replace", False)
-
-        # Kick off task to Do the Deed
-        task = dispatch(
-            tasks.upload_comps,
-            exclusive_resources=[repository] if repository else [],
-            args=([str(temp_file.pk), repo_pk, replace]),
-            kwargs={},
-        )
-        return OperationPostponedResponse(task, request)
-
-
-class RpmAlternateContentSourceViewSet(AlternateContentSourceViewSet, RolesMixin):
-    """
-    ViewSet for ACS.
-    """
-
-    endpoint_name = "rpm"
-    queryset = RpmAlternateContentSource.objects.all()
-    serializer_class = RpmAlternateContentSourceSerializer
-    queryset_filtering_required_permission = "rpm.view_rpmalternatecontentsource"
-
-    DEFAULT_ACCESS_POLICY = {
-        "statements": [
-            {
-                "action": ["list"],
-                "principal": ["authenticated"],
-                "effect": "allow",
-            },
-            {
-                "action": ["retrieve"],
-                "principal": "authenticated",
-                "effect": "allow",
-                "condition": "has_model_or_obj_perms:rpm.view_rpmalternatecontentsource",
-            },
-            {
-                "action": ["create"],
-                "principal": "authenticated",
-                "effect": "allow",
-                "condition": [
-                    "has_remote_param_model_or_obj_perms:rpm.view_rpmremote",
-                    "has_model_perms:rpm.add_rpmalternatecontentsource",
-                ],
-            },
-            {
-                "action": ["refresh"],
-                "principal": "authenticated",
-                "effect": "allow",
-                "condition": [
-                    "has_model_or_obj_perms:rpm.view_rpmalternatecontentsource",
-                    "has_model_perms:rpm.refresh_rpmalternatecontentsource",
-                ],
-            },
-            {
-                "action": ["update", "partial_update"],
-                "principal": "authenticated",
-                "effect": "allow",
-                "condition": [
-                    "has_model_or_obj_perms:rpm.change_rpmalternatecontentsource",
-                    "has_model_or_obj_perms:rpm.view_rpmalternatecontentsource",
-                    "has_remote_param_model_or_obj_perms:rpm.view_rpmremote",
-                ],
-            },
-            {
-                "action": ["destroy"],
-                "principal": "authenticated",
-                "effect": "allow",
-                "condition": [
-                    "has_model_or_obj_perms:rpm.delete_rpmalternatecontentsource",
-                    "has_model_or_obj_perms:rpm.view_rpmalternatecontentsource",
-                ],
-            },
-            {
-                "action": ["list_roles", "add_role", "remove_role"],
-                "principal": "authenticated",
-                "effect": "allow",
-                "condition": "has_model_or_obj_perms:rpm.manage_roles_rpmalternatecontentsource",
-            },
-        ],
-        "creation_hooks": [
-            {
-                "function": "add_roles_for_object_creator",
-                "parameters": {"roles": "rpm.rpmalternatecontentsource_owner"},
-            }
-        ],
-    }
-
-    LOCKED_ROLES = {
-        "rpm.rpmalternatecontentsource_owner": [
-            "rpm.change_rpmalternatecontentsource",
-            "rpm.delete_rpmalternatecontentsource",
-            "rpm.manage_roles_rpmalternatecontentsource",
-            "rpm.refresh_rpmalternatecontentsource",
-            "rpm.view_rpmalternatecontentsource",
-        ],
-        "rpm.rpmalternatecontentsource_creator": [
-            "rpm.add_rpmalternatecontentsource",
-        ],
-        "rpm.rpmalternatecontentsource_viewer": [
-            "rpm.view_rpmalternatecontentsource",
-        ],
-    }
-
-    @extend_schema(
-        description="Trigger an asynchronous task to create Alternate Content Source content.",
-        responses={202: TaskGroupOperationResponseSerializer},
-    )
-    @action(methods=["post"], detail=True)
-    def refresh(self, request, pk):
-        """
-        Refresh ACS metadata.
-        """
-        acs = get_object_or_404(RpmAlternateContentSource, pk=pk)
-        acs_paths = AlternateContentSourcePath.objects.filter(alternate_content_source=pk)
-        task_group = TaskGroup.objects.create(
-            description=f"Refreshing {acs_paths.count()} alternate content source path(s)."
-        )
-
-        # Get required defaults for sync operation
-        optimize = RpmRepositorySyncURLSerializer().data["optimize"]
-        skip_types = RpmRepositorySyncURLSerializer().data["skip_types"]
-
-        for acs_path in acs_paths:
-            # Create or get repository for the path
-            repo_data = {
-                "name": f"{acs.name}--{acs_path.pk}--repository",
-                "retain_repo_versions": 1,
-                "user_hidden": True,
-            }
-            repo, created = RpmRepository.objects.get_or_create(**repo_data)
-            if created:
-                acs_path.repository = repo
-                acs_path.save()
-            acs_url = (
-                os.path.join(acs.remote.url, acs_path.path) if acs_path.path else acs.remote.url
-            )
-
-            # Dispatching ACS path to own task and assign it to common TaskGroup
-            dispatch(
-                tasks.synchronize,
-                shared_resources=[acs.remote, acs],
-                task_group=task_group,
-                kwargs={
-                    "remote_pk": str(acs.remote.pk),
-                    "repository_pk": str(acs_path.repository.pk),
-                    "sync_policy": SYNC_POLICIES.MIRROR_CONTENT_ONLY,
-                    "skip_types": skip_types,
-                    "optimize": optimize,
-                    "url": acs_url,
-                },
-            )
-
-        # Update TaskGroup that all child task are dispatched
-        task_group.finish()
-
-        acs.last_refreshed = now()
-        acs.save()
-
-        return TaskGroupOperationResponse(task_group, request)
